fix(parquet): address Arrow and loader export feedback
