+++ modules/parquet/src/bundled.ts
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-export type {ParquetLoaderOptions} from './parquet-loader';
+export type {ParquetLoaderOptions} from './parquet-loader-types';
 export type {GeoParquetLoaderOptions} from './geoparquet-loader';
-export {ParquetWASMLoaderWithParser as ParquetLoader} from './parquet-wasm-loader-with-parser';
+export {ParquetLoaderWithParser as ParquetLoader} from './parquet-loader';
 export {GeoParquetLoaderWithParser as GeoParquetLoader} from './geoparquet-loader-with-parser';
-export {ParquetLoaderWithParser as ParquetJSLoader} from './parquet-loader-with-parser';
+export {ParquetJSLoaderWithParser as ParquetJSLoader} from './parquet-js-loader';
 export {
   ParquetSourceLoader,
   ParquetSource,
