+++ modules/parquet/src/index.ts
@@ -7,9 +7,9 @@
 export {ParquetFormat} from './parquet-format';
 
 export type {ParquetLoaderOptions, ParquetJSLoaderOptions} from './parquet-loader-options';
-export {ParquetLoader} from './parquet-loader';
+export {ParquetLoader} from './parquet-loader-types';
 export {GeoParquetLoader} from './geoparquet-loader';
-export {ParquetJSLoader} from './parquet-js-loader';
+export {ParquetJSLoader} from './parquet-js-loader-types';
 
 export {ParquetSourceLoader} from './parquet-source-loader-types';
 export type {
