+++ modules/parquet/src/unbundled.ts
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-export type {ParquetLoaderOptions} from './parquet-loader';
+export type {ParquetLoaderOptions} from './parquet-loader-types';
 export type {GeoParquetLoaderOptions} from './geoparquet-loader';
-export {ParquetLoader} from './parquet-loader';
+export {ParquetLoader} from './parquet-loader-types';
 export {GeoParquetLoader} from './geoparquet-loader';
-export {ParquetJSLoader} from './parquet-js-loader';
+export {ParquetJSLoader} from './parquet-js-loader-types';
 export {
   ParquetSourceLoader,
   ParquetSource,
