+++ modules/parquet/test/parquet-loader.spec.ts
@@ -8,9 +8,9 @@ import test from 'test/utils/vitest-tape';
 import {validateLoader} from 'test/common/conformance';
 
 import {ParquetJSLoader, ParquetLoader} from '@loaders.gl/parquet';
+import {ParquetJSLoaderWithParser} from '@loaders.gl/parquet/parquet-js-loader';
+import {ParquetLoaderWithParser} from '@loaders.gl/parquet/parquet-loader';
 import {isBrowser, load, preload, setLoaderOptions} from '@loaders.gl/core';
-import {ParquetLoaderWithParser} from '../src/parquet-loader-with-parser';
-import {ParquetWASMLoaderWithParser} from '../src/parquet-wasm-loader-with-parser';
 
 import {SUPPORTED_FILES, UNSUPPORTED_FILES, ENCRYPTED_FILES, BAD_FILES} from './data/files';
 import {
@@ -52,15 +52,15 @@ test('ParquetJSLoader#loader objects', (t) => {
 test('Parquet loaders preload explicit parser implementations', async (t) => {
   t.equal(
     await preload(ParquetLoader),
-    ParquetWASMLoaderWithParser,
+    ParquetLoaderWithParser,
     'primary ParquetLoader resolves the WASM parser'
   );
   t.equal(
     await preload(ParquetJSLoader),
-    ParquetLoaderWithParser,
+    ParquetJSLoaderWithParser,
     'fallback ParquetJSLoader resolves the TypeScript parser'
   );
-  t.equal(ParquetLoaderWithParser.id, ParquetJSLoader.id, 'fallback parser preserves loader id');
+  t.equal(ParquetJSLoaderWithParser.id, ParquetJSLoader.id, 'fallback parser preserves loader id');
   t.equal(ParquetJSLoader.worker, false, 'fallback parser stays on the main thread');
   t.end();
 });
@@ -87,6 +87,26 @@ test('ParquetJSLoader#load supports arrow-table shape', async (t) => {
   t.equal(table.shape, 'arrow-table');
   if (table.shape === 'arrow-table') {
     t.equal(table.data.numRows, 2);
+  }
+  t.end();
+});
+
+test('ParquetJSLoader#load arrow-table preserves schema for empty results', async (t) => {
+  const url = '@loaders.gl/parquet/test/data/apache/good/alltypes_dictionary.parquet';
+  const table = await load(url, ParquetJSLoader, {
+    core: {worker: false},
+    parquet: {shape: 'arrow-table', limit: 0}
+  });
+
+  t.equal(table.shape, 'arrow-table');
+  if (table.shape === 'arrow-table') {
+    t.equal(table.data.numRows, 0);
+    t.ok(table.schema.fields.length > 0, 'loaders.gl schema retains the file fields');
+    t.deepEqual(
+      table.data.schema.fields.map((field) => field.name),
+      table.schema.fields.map((field) => field.name),
+      'Arrow schema retains the same file fields'
+    );
   }
   t.end();
 });
@@ -188,6 +208,26 @@ test('ParquetJSLoader#load nested_lists file', async (t) => {
   t.end();
 });
 
+test('ParquetJSLoader#load nested_lists file as an Arrow table', async (t) => {
+  const url = '@loaders.gl/parquet/test/data/apache/good/nested_lists.snappy.parquet';
+  const table = await load(url, ParquetJSLoader, {
+    core: {worker: false},
+    parquet: {shape: 'arrow-table'}
+  });
+
+  t.equal(table.shape, 'arrow-table');
+  if (table.shape === 'arrow-table') {
+    const firstRow = JSON.parse(JSON.stringify(table.data.get(0)?.toJSON())) as {
+      a: {list: Array<{element: {list: unknown[]}}>};
+    };
+    t.equal(table.data.numRows, 3);
+    t.ok(JSON.stringify(table.schema).includes('"type":"list"'), 'schema contains Arrow lists');
+    t.equal(firstRow.a.list.length, 2, 'outer repeated group is preserved');
+    t.equal(firstRow.a.list[0].element.list.length, 2, 'nested repeated group is preserved');
+  }
+  t.end();
+});
+
 test('ParquetJSLoader#load nested_maps file', async (t) => {
   const url = '@loaders.gl/parquet/test/data/apache/good/nested_maps.snappy.parquet';
   const table = await load(url, ParquetJSLoader, getParquetLoaderOptions(url));
@@ -263,6 +303,26 @@ test('ParquetJSLoader#load repeated_no_annotation file', async (t) => {
   if (table.shape === 'object-row-table') {
     t.equal(table.data.length, 6);
     t.deepEqual(table.data, REPEATED_NO_ANNOTATION_EXPECTED);
+  }
+  t.end();
+});
+
+test('ParquetJSLoader#load repeated_no_annotation file as an Arrow table', async (t) => {
+  const url = '@loaders.gl/parquet/test/data/apache/good/repeated_no_annotation.parquet';
+  const table = await load(url, ParquetJSLoader, {
+    core: {worker: false},
+    parquet: {shape: 'arrow-table'}
+  });
+
+  t.equal(table.shape, 'arrow-table');
+  if (table.shape === 'arrow-table') {
+    const lastRow = JSON.parse(JSON.stringify(table.data.get(5)?.toJSON())) as {
+      phoneNumbers: {phone: Array<{number: number; kind: string | null}>};
+    };
+    t.equal(table.data.numRows, 6);
+    t.ok(JSON.stringify(table.schema).includes('"type":"list"'), 'schema contains Arrow lists');
+    t.equal(lastRow.phoneNumbers.phone.length, 3, 'all repeated structs are retained');
+    t.equal(lastRow.phoneNumbers.phone[2].kind, 'mobile', 'repeated struct values are retained');
   }
   t.end();
 });
